feat: add google sign in on web.
Comes with a complete example implementation.

You are required to setup a webpage to send an event back to the main flutter page for the webpart to work.
There is a ready to use backend widget for this purpose. Other than this fact the implementation are completely
backwards compatible with the old API.

Comes with