fix: Allow users on AWS to set/upgrade dart version.
This removes the installation of dart from launch template and install during code deploy. The purpose for this:
1. launch template only loads once an old instance is deleted, which means for upgrading dart, users need to delete the old machine and launch a new one which causes significant downtime
2. this allows multiple versions of dart to be installed on the instance, but only the specified one will be used to run code

for users using prior version of serverpod cli, we have commented out a section of the code in `install_dependencies`. users can uncomment the code and upgrade their dart version on running instance, the downtime will be approximately the same as rolling out a code release