refactor: rename getter methods to be intentionally opaque
The idea is that these should be able to validate an arbitrary amount of things and own all business logic for endpoints. Therefore it is clearer and more future proof to be intentionally opaque. Together with the new return type `MethodCallContext`, it should be clear to the user of this function what the intent is.