refactor: guard for experimental features
A singleton was added which can be instantiated and holds a list of experimental features.

the multiOption on the argParser was moved from the Generate command to ServerpodCommandRunner

another option was added to pass --experimental-features=all which enables all experimental features. This is used in migration_test_utils.dart and util/generate_all