feat: Enable adding custom data to diagnostic events (#3363)
This adds the ability to include custom data when submitting diagnostic events.

It is intended to be used by application (user) code, not framework code. The content to include and how to handle it in the event handlers is up to the user.

Closes #3332 

The solution differs slightly from the suggestion in the issue: Rather than extending the context, this is included as a key/value map in the event itself. This makes sense since the context types are framework-managed and this is an entirely user-owned data payload for individual events.

## Pre-launch Checklist

- [x] I read the [Contribute](https://docs.serverpod.dev/contribute) page and followed the process outlined there for submitting PRs.
- [x] This update contains only one single feature or bug fix and nothing else. (If you are submitting multiple fixes, please make multiple PRs.)
- [x] I read and followed the [Dart Style Guide](https://dart.dev/guides/language/effective-dart/style) and formatted the code with [dart format](https://dart.dev/tools/dart-format).
- [x] I listed at least one issue that this PR fixes in the description above.
- [x] I updated/added relevant documentation (doc comments with `///`), and made sure that the documentation follows the same style as other Serverpod documentation. I checked spelling and grammar.
- [x] I added new tests to check the change I am making.
- [x] All existing and new tests are passing.
- [x] Any breaking changes are documented below.

If you need help, consider asking for advice on the [discussion board](https://github.com/serverpod/serverpod/discussions).

## Breaking changes

No, does not change any existing API. Adds an optional initialization field to the diagnostic event constructor.